compile_package_jars {sparklyr}	R Documentation

Compile Scala sources into a Java Archive (jar)

Description

Compile the scala source files contained within an R package into a Java Archive (jar) file that can be loaded and used within a Spark environment.

Usage

compile_package_jars(..., spec = NULL)

Arguments

...
Optional compilation specifications, as generated by spark_compilation_spec. When no arguments are passed, spark_default_compilation_spec is used instead.

spec
An optional list of compilation specifications. When set, this option takes precedence over arguments passed to ....

[Package sparklyr version 1.0.5 Index]