connection_spark_shinyapp {sparklyr}	R Documentation

A Shiny app that can be used to construct a spark_connect statement

Description

A Shiny app that can be used to construct a spark_connect statement

Usage

connection_spark_shinyapp()

[Package sparklyr version 0.8.0 Index]