Feature Tranformation -- StringIndexer (Estimator)

A label indexer that maps a string column of labels to an ML column of label indices. If the input column is numeric, we cast it to string and index the string values. The indices are in [0, numLabels), ordered by label frequencies. So the most frequent label gets index 0. This function is the inverse of ft_index_to_string.

ft_string_indexer(x, input_col, output_col, handle_invalid = "error",
  dataset = NULL, uid = random_string("string_indexer_"), ...)

Arguments

x
A spark_connection, ml_pipeline, or a tbl_spark.

input_col
The name of the input column.

output_col
The name of the output column.

handle_invalid
Param for how to handle invalid entries. Options are 'skip' (filter out rows with invalid values), 'error' (throw an error), or 'keep' (keep invalid values in a special additional bucket). Default: "error"

dataset
(Optional) A tbl_spark. If provided, eagerly fit the (estimator) feature "transformer" against dataset. See details.

uid
A character string used to uniquely identify the feature transformer.

...
Optional arguments; currently unused.

Value

The object returned depends on the class of x.

spark_connection: When x is a spark_connection, the function returns a ml_transformer, a ml_estimator, or one of their subclasses. The object contains a pointer to a Spark Transformer or Estimator object and can be used to compose Pipeline objects.

ml_pipeline: When x is a ml_pipeline, the function returns a ml_pipeline with the transformer or estimator appended to the pipline.

tbl_spark: When x is a tbl_spark, a transformer is constructed then immediately applied to the input tbl_spark, returning a tbl_spark

Details

When dataset is provided for an estimator transformer, the function internally calls ml_fit() against dataset. Hence, the methods for spark_connection and ml_pipeline will then return a ml_transformer and a ml_pipeline with a ml_transformer appended, respectively. When x is a tbl_spark, the estimator will be fit against dataset before transforming x.

When dataset is not specified, the constructor returns a ml_estimator, and, in the case where x is a tbl_spark, the estimator fits against x then to obtain a transformer, which is then immediately used to transform x.

See also

See http://spark.apache.org/docs/latest/ml-features.html for more information on the set of transformations available for DataFrame columns in Spark. ft_index_to_string

Other feature transformers: ft_binarizer, ft_bucketizer, ft_count_vectorizer, ft_dct, ft_elementwise_product, ft_hashing_tf, ft_index_to_string, ft_one_hot_encoder, ft_pca, ft_quantile_discretizer, ft_r_formula, ft_regex_tokenizer, ft_sql_transformer, ft_stop_words_remover, ft_tokenizer, ft_vector_assembler