ml-model-constructors {sparklyr}	R Documentation

Constructors for 'ml_model' Objects

Description

Functions for developers writing extensions for Spark ML. These functions are constructors for 'ml_model' objects that are returned when using the formula interface.

Usage

new_ml_model_prediction(pipeline_model, formula, dataset, label_col,
  features_col, ..., class = character())

new_ml_model(pipeline_model, formula, dataset, ..., class = character())

new_ml_model_classification(pipeline_model, formula, dataset, label_col,
  features_col, predicted_label_col, ..., class = character())

new_ml_model_regression(pipeline_model, formula, dataset, label_col,
  features_col, ..., class = character())

new_ml_model_clustering(pipeline_model, formula, dataset, features_col,
  ..., class = character())

ml_supervised_pipeline(predictor, dataset, formula, features_col,
  label_col)

ml_clustering_pipeline(predictor, dataset, formula, features_col)

ml_construct_model_supervised(constructor, predictor, formula, dataset,
  features_col, label_col, ...)

ml_construct_model_clustering(constructor, predictor, formula, dataset,
  features_col, ...)

Arguments

pipeline_model
The pipeline model object returned by 'ml_supervised_pipeline()'.

formula
The formula used for data preprocessing

dataset
The training dataset.

label_col
Label column name. The column should be a numeric column. Usually this column is output by ft_r_formula.

features_col
Features column name, as a length-one character vector. The column should be single vector column of numeric values. Usually this column is output by ft_r_formula.

class
Name of the subclass.

predictor
The pipeline stage corresponding to the ML algorithm.

constructor
The constructor function for the 'ml_model'.

[Package sparklyr version 1.0.4 Index]