Spark ML -- Alternating Least Squares (ALS) matrix factorization.

Perform alternating least squares matrix factorization on a Spark DataFrame.

ml_als_factorization(x, rating.column = "rating", user.column = "user",
  item.column = "item", rank = 10L, regularization.parameter = 0.1,
  implicit.preferences = FALSE, alpha = 1, nonnegative = FALSE,
  iter.max = 10L, ml.options = ml_options(), ...)

Arguments

x
An object coercable to a Spark DataFrame (typically, a tbl_spark).

rating.column
The name of the column containing ratings.

user.column
The name of the column containing user IDs.

item.column
The name of the column containing item IDs.

rank
Rank of the factorization.

regularization.parameter
The regularization parameter.

implicit.preferences
Use implicit preference.

alpha
The parameter in the implicit preference formulation.

nonnegative
Use nonnegative constraints for least squares.

iter.max
The maximum number of iterations to use.

ml.options
Optional arguments, used to affect the model generated. See ml_options for more details.

...
Optional arguments. The data argument can be used to specify the data to be used when x is a formula; this allows calls of the form ml_linear_regression(y ~ x, data = tbl), and is especially useful in conjunction with do.

See also

Other Spark ML routines: ml_decision_tree, ml_generalized_linear_regression, ml_gradient_boosted_trees, ml_kmeans, ml_lda, ml_linear_regression, ml_logistic_regression, ml_multilayer_perceptron, ml_naive_bayes, ml_one_vs_rest, ml_pca, ml_random_forest, ml_survival_regression