ml_corr {sparklyr}	R Documentation

Compute correlation matrix

Description

Compute correlation matrix

Usage

ml_corr(x, columns = NULL, method = c("pearson", "spearman"))

Arguments

x
A tbl_spark.

columns
The names of the columns to calculate correlations of. If only one column is specified, it must be a vector column (for example, assembled using ft_vector_assember()).

method
The method to use, either "pearson" or "spearman".

Value

A correlation matrix organized as a data frame.

[Package sparklyr version 0.8.3 Index]