Default stop words

Loads the default stop words for the given language.

ml_default_stop_words(sc, language = c("danish", "dutch", "english",
  "finnish", "french", "german", "hungarian", "italian", "norwegian",
  "portuguese", "russian", "spanish", "swedish", "turkish"), ...)

Arguments

sc
A spark_connection

language
A character string.

...
Optional arguments; currently unused.

Value

A list of stop words.

Details

Supported languages: danish, dutch, english, finnish, french, german, hungarian, italian, norwegian, portuguese, russian, spanish, swedish, turkish. See http://anoncvs.postgresql.org/cvsweb.cgi/pgsql/src/backend/snowball/stopwords/ for more details

See also

ft_stop_words_remover