ml_glm_tidiers {sparklyr}	R Documentation

Tidying methods for Spark ML linear models

Description

These methods summarize the results of Spark ML models into tidy forms.

Usage

## S3 method for class 'ml_model_generalized_linear_regression'
tidy(x, exponentiate = FALSE,
  ...)

## S3 method for class 'ml_model_linear_regression'
tidy(x, ...)

## S3 method for class 'ml_model_generalized_linear_regression'
augment(x, newdata = NULL,
  type.residuals = c("working", "deviance", "pearson", "response"), ...)

## S3 method for class 'ml_model_linear_regression'
augment(x, newdata = NULL,
  type.residuals = c("working", "deviance", "pearson", "response"), ...)

## S3 method for class 'ml_model_generalized_linear_regression'
glance(x, ...)

## S3 method for class 'ml_model_linear_regression'
glance(x, ...)

Arguments

x
a Spark ML model.

exponentiate
For GLM, whether to exponentiate the coefficient estimates (typical for logistic regression.)

...
extra arguments (not used.)

newdata
a tbl_spark of new data to use for prediction.

type.residuals
type of residuals, defaults to "working". Must be set to "working" when newdata is supplied.

Details

The residuals attached by augment are of type "working" by default, which is different from the default of "deviance" for residuals() or sdf_residuals().

[Package sparklyr version 0.8.0 Index]