sdf_debug_string {sparklyr}	R Documentation

Debug Info for Spark DataFrame

Description

Prints plan of execution to generate x. This plan will, among other things, show the number of partitions in parenthesis at the far left and indicate stages using indentation.

Usage

sdf_debug_string(x, print = TRUE)

Arguments

x
An R object wrapping, or containing, a Spark DataFrame.

print
Print debug information?

[Package sparklyr version 1.0.5 Index]