sdf_describe {sparklyr}	R Documentation

Compute summary statistics for columns of a data frame

Description

Compute summary statistics for columns of a data frame

Usage

sdf_describe(x, cols = colnames(x))

Arguments

x
An object coercible to a Spark DataFrame

cols
Columns to compute statistics for, given as a character vector

[Package sparklyr version 0.8.3 Index]