sdf_schema {sparklyr}	R Documentation

Read the Schema of a Spark DataFrame

Description

Read the schema of a Spark DataFrame.

Usage

sdf_schema(x)

Arguments

x
A spark_connection, ml_pipeline, or a tbl_spark.

Details

The type column returned gives the string representation of the underlying Spark type for that column; for example, a vector of numeric values would be returned with the type "DoubleType". Please see the Spark Scala API Documentation for information on what types are available and exposed by Spark.

Value

An R list, with each list element describing the name and type of a column.

[Package sparklyr version 0.8.2 Index]