sdf_seq {sparklyr}	R Documentation

Create DataFrame for Range

Description

Creates a DataFrame for the given range

Usage

sdf_seq(sc, from = 1L, to = 1L, by = 1L, repartition = NULL)

Arguments

sc
The associated Spark connection.

from, to
The start and end to use as a range

by
The increment of the sequence.

repartition
The number of partitions to use when distributing the data across the Spark cluster.

[Package sparklyr version 0.8.2 Index]