spark_dataframe {sparklyr}	R Documentation

Retrieve a Spark DataFrame

Description

This S3 generic is used to access a Spark DataFrame object (as a Java object reference) from an R object.

Usage

spark_dataframe(x, ...)

Arguments

x
An R object wrapping, or containing, a Spark DataFrame.

...
Optional arguments; currently unused.

Value

A spark_jobj representing a Java object reference to a Spark DataFrame.

[Package sparklyr version 0.8.3 Index]