spark_default_version {sparklyr}	R Documentation

determine the version that will be used by default if version is NULL

Description

determine the version that will be used by default if version is NULL

Usage

spark_default_version()

[Package sparklyr version 0.7.0 Index]