spark_jobj {sparklyr}	R Documentation

Retrieve a Spark JVM Object Reference

Description

This S3 generic is used for accessing the underlying Java Virtual Machine (JVM) Spark objects associated with R objects. These objects act as references to Spark objects living in the JVM. Methods on these objects can be called with the invoke family of functions.

Usage

spark_jobj(x, ...)

Arguments

x
An R object containing, or wrapping, a spark_jobj.

...
Optional arguments; currently unused.

See Also

invoke, for calling methods on Java object references.

[Package sparklyr version 0.8.3 Index]