spark_log {sparklyr}	R Documentation

View Entries in the Spark Log

Description

View the most recent entries in the Spark log. This can be useful when inspecting output / errors produced by Spark during the invocation of various commands.

Usage

spark_log(sc, n = 100, filter = NULL, ...)

Arguments

sc
A spark_connection.

n
The max number of log entries to retrieve. Use NULL to retrieve all entries within the log.

filter
Character string to filter log entries.

...
Optional arguments; currently unused.

[Package sparklyr version 1.0.4 Index]