spark_versions {sparklyr}	R Documentation

Retrieves a dataframe available Spark versions that van be installed.

Description

Retrieves a dataframe available Spark versions that van be installed.

Usage

spark_versions(latest = TRUE)

Arguments

latest
Check for latest version?

[Package sparklyr version 0.8.3 Index]