Write a Spark DataFrame to a Text file

Serialize a Spark DataFrame to the plain text format.

spark_write_text(x, path, mode = NULL, options = list(),
  partition_by = NULL, ...)

Arguments

x
A Spark DataFrame or dplyr operation

path
The path to the file. Needs to be accessible from the cluster. Supports the "hdfs://", "s3n://" and "file://" protocols.

mode
Specifies the behavior when data or table already exists.

options
A list of strings with additional options.

partition_by
Partitions the output by the given columns on the file system.

...
Optional arguments; currently unused.

See also

Other Spark serialization routines: spark_load_table, spark_read_csv, spark_read_jdbc, spark_read_json, spark_read_parquet, spark_read_source, spark_read_table, spark_read_text, spark_save_table, spark_write_csv, spark_write_jdbc, spark_write_json, spark_write_parquet, spark_write_source, spark_write_table