Read Delta Stream

Reads a Delta Lake table as a Spark dataframe stream.

stream_read_delta(sc, path, name = NULL, options = list(), ...)

Arguments

sc
A spark_connection.

path
The path to the file. Needs to be accessible from the cluster. Supports the "hdfs://", "s3a://" and "file://" protocols.

name
The name to assign to the newly generated stream.

options
A list of strings with additional options.

...
Optional arguments; currently unused.

Details

Please note that Delta Lake requires installing the appropriate package by setting the packages parameter to "delta" in spark_connect()

See also

Examples

if (FALSE) {

library(sparklyr)
sc <- spark_connect(master = "local", version = "2.4.0", packages = "delta")

sdf_len(sc, 5) %>% spark_write_delta(path = "delta-test")

stream <- stream_read_delta(sc, "delta-test") %>%
  stream_write_json("json-out")

stream_stop(stream)
}