stream_read_kafka {sparklyr}	R Documentation

Read Kafka Stream

Description

Reads a Kafka stream as a Spark dataframe stream.

Usage

stream_read_kafka(sc, name = NULL, options = list(), ...)

Arguments

sc
A spark_connection.

name
The name to assign to the newly generated stream.

options
A list of strings with additional options.

...
Optional arguments; currently unused.

Details

Please note that Kafka requires installing the appropriate package by conneting with a config setting where sparklyr.shell.packages is set to, for Spark 2.3.2, "org.apache.spark:spark-sql-kafka-0-10_2.11:2.3.2".

See Also

Other Spark stream serialization: stream_read_csv, stream_read_json, stream_read_orc, stream_read_parquet, stream_read_scoket, stream_read_text, stream_write_console, stream_write_csv, stream_write_json, stream_write_kafka, stream_write_memory, stream_write_orc, stream_write_parquet, stream_write_text

Examples

## Not run: 

config <- spark_config()

# The following package is dependent to Spark version, for Spark 2.3.2:
config$sparklyr.shell.packages <- "org.apache.spark:spark-sql-kafka-0-10_2.11:2.3.2"

sc <- spark_connect(master = "local", config = config)

read_options <- list(kafka.bootstrap.servers = "localhost:9092", subscribe = "topic1")
write_options <- list(kafka.bootstrap.servers = "localhost:9092", topic = "topic2")

stream <- stream_read_kafka(sc, options = read_options) %>%
  stream_write_kafka(options = write_options)

stream_stop(stream)

## End(Not run)

[Package sparklyr version 1.0.4 Index]