stream_read_scoket {sparklyr}	R Documentation

Read Socket Stream

Description

Reads a Socket stream as a Spark dataframe stream.

Usage

stream_read_scoket(sc, name = NULL, columns = NULL, options = list(),
  ...)

Arguments

sc
A spark_connection.

name
The name to assign to the newly generated stream.

columns
A vector of column names or a named vector of column types.

options
A list of strings with additional options.

...
Optional arguments; currently unused.

See Also

Other Spark stream serialization: stream_read_csv, stream_read_json, stream_read_kafka, stream_read_orc, stream_read_parquet, stream_read_text, stream_write_console, stream_write_csv, stream_write_json, stream_write_kafka, stream_write_memory, stream_write_orc, stream_write_parquet, stream_write_text

Examples

## Not run: 

sc <- spark_connect(master = "local")

# Start socket server from terminal, example: nc -lk 9999
stream <- stream_read_scoket(sc, options = list(host = "localhost", port = 9999))
stream

## End(Not run)

[Package sparklyr version 1.0.4 Index]