stream_trigger_continuous {sparklyr}	R Documentation

Spark Stream Continuous Trigger

Description

Creates a Spark structured streaming trigger to execute continuously. This mode is the most performant but not all operations are supported.

Usage

stream_trigger_continuous(checkpoint = 5000)

Arguments

checkpoint
The checkpoint interval specified in milliseconds.

See Also

stream_trigger_interval

[Package sparklyr version 1.0.4 Index]