Write a ORC Stream

Writes a Spark dataframe stream into an ORC stream.

stream_write_orc(
  x,
  path,
  mode = c("append", "complete", "update"),
  trigger = stream_trigger_interval(),
  checkpoint = file.path(path, "checkpoints", random_string("")),
  options = list(),
  ...
)

Arguments

x
A Spark DataFrame or dplyr operation

path
The destination path. Needs to be accessible from the cluster. Supports the "hdfs://", "s3a://" and "file://" protocols.

mode
Specifies how data is written to a streaming sink. Valid values are "append", "complete" or "update".

trigger
The trigger for the stream query, defaults to micro-batches runnnig every 5 seconds. See stream_trigger_interval and stream_trigger_continuous.

checkpoint
The location where the system will write all the checkpoint information to guarantee end-to-end fault-tolerance.

options
A list of strings with additional options.

...
Optional arguments; currently unused.

See also

Examples

if (FALSE) {

sc <- spark_connect(master = "local")

sdf_len(sc, 10) %>% spark_write_orc("orc-in")

stream <- stream_read_orc(sc, "orc-in") %>% stream_write_orc("orc-out")

stream_stop(stream)

}