stream_render {sparklyr}	R Documentation

Render Stream

Description

Collects streaming statistics to render the stream as an 'htmlwidget'.

Usage

stream_render(stream, collect = 10, ...)

Arguments

stream
The stream to render

collect
The interval in seconds to collect data before rendering the 'htmlwidget'.

...
Additional optional arguments.

#' @examples

library(sparklyr)

dir.create("iris-in") write.csv(iris, "iris-in/iris.csv", row.names = FALSE)

stream <- stream_read_csv("iris-in/") stream_write_csv("iris-out/")

stream_render(stream) stream_stop(stream)

[Package sparklyr version 0.8.4.9004 Index]